requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
appdirs==1.4.4
beautifulsoup4==4.9.3
bs4==0.0.2
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==2.0.12
click==8.3.1
cryptography==46.0.6
cssselect==1.4.0
cuda-bindings==13.2.0
cuda-pathfinder==1.5.1
cuda-toolkit==13.0.2
dateparser==1.4.0
distro==1.9.0
dnspython==1.16.0
duckduckgo_search==8.1.1
fake-useragent==2.2.0
feedfinder2==0.0.4
feedparser==6.0.12
filelock==3.25.2
fsspec==2026.3.0
gnews==0.3.6
google-ai-generativelanguage==0.6.15
google-api-core==2.30.1
google-api-python-client==2.193.0
google-auth==2.49.1
google-auth-httplib2==0.3.1
google-generativeai==0.8.4
googleapis-common-protos==1.73.1
googlenewsdecoder==0.1.7
grpcio==1.80.0
grpcio-status==1.71.2
h11==0.16.0
hf-xet==1.4.3
httpcore==1.0.9
httplib2==0.31.2
httpx==0.28.1
huggingface_hub==1.9.0
idna==3.11
importlib_metadata==9.0.0
jieba3k==0.35.1
Jinja2==3.1.6
jiter==0.13.0
joblib==1.5.3
lxml==6.0.2
lxml_html_clean==0.4.4
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
mpmath==1.3.0
networkx==3.6.1
newspaper3k==0.2.8
nltk==3.9.4
numpy==2.4.4
nvidia-cublas==13.1.0.3
nvidia-cuda-cupti==13.0.85
nvidia-cuda-nvrtc==13.0.88
nvidia-cuda-runtime==13.0.96
nvidia-cudnn-cu13==9.19.0.56
nvidia-cufft==12.0.0.61
nvidia-cufile==1.15.1.6
nvidia-curand==10.4.0.35
nvidia-cusolver==12.0.4.66
nvidia-cusparse==12.6.3.3
nvidia-cusparselt-cu13==0.8.0
nvidia-nccl-cu13==2.28.9
nvidia-nvjitlink==13.0.88
nvidia-nvshmem-cu13==3.4.5
nvidia-nvtx==13.0.85
openai==2.30.0
packaging==26.0
parse==1.21.1
pillow==12.2.0
primp==1.2.1
proto-plus==1.27.2
protobuf==5.29.6
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
pyee==11.1.1
Pygments==2.20.0
pymongo==3.12.3
pyparsing==3.3.2
pyppeteer==2.0.0
pyquery==2.0.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==0.19.2
pytz==2026.1.post1
PyYAML==6.0.3
regex==2026.3.32
requests==2.33.1
requests-file==3.0.1
requests-html==0.10.0
rich==14.3.3
safetensors==0.7.0
Sastrawi==1.0.1
selectolax==0.4.7
setuptools==81.0.0
sgmllib3k==1.0.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.8.3
sympy==1.14.0
tinysegmenter==0.3
tldextract==5.3.1
tokenizers==0.22.2
torch==2.11.0
tqdm==4.67.3
transformers==5.5.0
triton==3.6.0
typer==0.24.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzlocal==5.3.1
uritemplate==4.2.0
urllib3==1.26.20
w3lib==2.4.1
websockets==10.4
zipp==3.23.0
